refactor(api): create mutations and query folders
- divide api calls based of their functionality
- move get request to queries
- move post/delete/put into mutations
