+++ apps/ui/src/api/mutations/useCreateApplication.ts
