+++ apps/ui/src/api/mutations/useEditApplication.ts
