+++ apps/ui/src/api/queries/useGetApplication.ts
