+++ apps/ui/src/api/queries/useGetApplicationList.ts
