+++ apps/ui/src/components/pages/application/SectionFooter.tsx
@@ -17,7 +17,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import useEditApplication from '@/api/useEditApplication';
+import useEditApplication from '@/api/mutations/useEditApplication';
 import { ApplicationOutletContext } from '@/global/types';
 import { ApplicationSectionRoutes } from '@/pages/AppRouter';
 import { useApplicationContext } from '@/providers/context/application/ApplicationContext';
